Update adc_config_t: change vref_millivolt type from uint8_t to uint16_t

+++ projects/01_sensor_fusion/drivers/adc_types.h
@@ -16,7 +16,7 @@ typedef enum
 typedef struct 
 {
     uint8_t channel;
-    uint8_t vref_millivolt;
+    uint16_t vref_millivolt;
     uint8_t oversample_count;
     bool as_millivolt;
 } adc_config_t;
